requirements.txt
beautifulsoup4==4.8.1
bitarray==1.0.1
blis==0.4.1
Bottleneck==1.2.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
fastai==1.0.58
fastprogress==0.1.21
Flask==1.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lshash3==0.0.8
MarkupSafe==1.1.1
matplotlib==3.1.1
murmurhash==1.0.2
numexpr==2.7.0
numpy==1.17.2
nvidia-ml-py3==7.352.0
opencv-contrib-python==4.1.1.26
packaging==19.2
pandas==0.25.1
pickleshare==0.7.5
Pillow==6.2.0
plac==0.9.6
preshed==3.0.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scipy==1.3.1
six==1.12.0
soupsieve==1.9.4
spacy==2.2.1
srsly==0.1.0
thinc==7.1.1
torch==1.3.0
torchvision==0.4.1
tqdm==4.36.1
urllib3==1.25.6
wasabi==0.2.2
Werkzeug==0.16.0
WTForms==2.2.1
